requirements.txt
grpcio
grpcio-tools
pip
protobuf==5.27.2
setuptools
python-dotenv
